+++ core/components/com_publications/admin/views/types/tmpl/display.php
@@ -32,18 +32,30 @@
 // No direct access
 defined('_HZEXEC_') or die();
 
-Toolbar::title(Lang::txt('COM_PUBLICATIONS_PUBLICATIONS') . ': ' . Lang::txt('COM_PUBLICATIONS_MASTER_TYPES'), 'addedit.png');
-Toolbar::addNew();
-Toolbar::editList();
-Toolbar::spacer();
-Toolbar::deleteList();
+$canDo = \Components\Publications\Helpers\Permissions::getActions('type');
 
+Toolbar::title(Lang::txt('COM_PUBLICATIONS_PUBLICATIONS') . ': ' . Lang::txt('COM_PUBLICATIONS_MASTER_TYPES'));
+if ($canDo->get('core.create'))
+{
+	Toolbar::addNew();
+}
+if ($canDo->get('core.edit'))
+{
+	Toolbar::editList();
+}
+if ($canDo->get('core.delete'))
+{
+	Toolbar::spacer();
+	Toolbar::deleteList();
+}
+Toolbar::divider();
+Toolbar::help('types');
 ?>
 <form action="<?php echo Route::url('index.php?option=' . $this->option . '&controller=' . $this->controller); ?>" method="post" name="adminForm" id="adminForm">
 	<table class="adminlist">
 		<thead>
 			<tr>
-				<th><input type="checkbox" name="toggle" value="" onclick="checkAll(<?php echo count( $this->rows );?>);" /></th>
+				<th><input type="checkbox" name="toggle" value="" onclick="checkAll(<?php echo count($this->rows);?>);" /></th>
 				<th class="priority-3"><?php echo Html::grid('sort', Lang::txt('COM_PUBLICATIONS_FIELD_ID'), 'id', @$this->filters['sort_Dir'], @$this->filters['sort'] ); ?></th>
 				<th><?php echo Html::grid('sort', Lang::txt('COM_PUBLICATIONS_FIELD_NAME'), 'type', @$this->filters['sort_Dir'], @$this->filters['sort'] ); ?></th>
 				<th class="priority-4"><?php echo Lang::txt('COM_PUBLICATIONS_FIELD_ALIAS'); ?></th>
@@ -64,30 +76,30 @@ Toolbar::deleteList();
 			</tr>
 		</tfoot>
 		<tbody>
-<?php
-$k = 0;
-for ($i=0, $n=count( $this->rows ); $i < $n; $i++)
-{
-	$row = &$this->rows[$i];
-	$sClass = $row->supporting == 1 ? 'on' : 'off';
-	$cClass = $row->contributable == 1 ? 'on' : 'off';
+		<?php
+		$k = 0;
+		for ($i=0, $n=count($this->rows); $i < $n; $i++)
+		{
+			$row = &$this->rows[$i];
+			$sClass = $row->supporting == 1 ? 'on' : 'off';
+			$cClass = $row->contributable == 1 ? 'on' : 'off';
 
-	// Determine whether master type is supported in current version of hub code
-	$aClass = 'off';
-	$active = 'off';
+			// Determine whether master type is supported in current version of hub code
+			$aClass = 'off';
+			$active = 'off';
 
-	// If we got a plugin - type is supported
-	if (Plugin::isEnabled('projects', $row->alias))
-	{
-		$aClass = 'on';
-		$active = 'on';
-	}
-?>
+			// If we got a plugin - type is supported
+			if (Plugin::isEnabled('projects', $row->alias))
+			{
+				$aClass = 'on';
+				$active = 'on';
+			}
+			?>
 			<tr class="<?php echo "row$k"; ?>">
-				<td class="narrow">
+				<td>
 					<input type="checkbox" name="id[]" id="cb<?php echo $i; ?>" value="<?php echo $row->id; ?>" onclick="isChecked(this.checked);" />
 				</td>
-				<td class="priority-3 narrow">
+				<td class="priority-3">
 					<?php echo $row->id; ?>
 				</td>
 				<td>
@@ -98,7 +110,7 @@ for ($i=0, $n=count( $this->rows ); $i < $n; $i++)
 				<td class="priority-4">
 					<span class="block faded"><?php echo $this->escape($row->alias); ?></span>
 				</td>
-				<td class="priority-3 centeralign narrow">
+				<td class="priority-3">
 					<span class="state <?php echo $cClass; ?>">
 						<span><?php echo Lang::txt($cClass); ?></span>
 					</span>
@@ -113,10 +125,10 @@ for ($i=0, $n=count( $this->rows ); $i < $n; $i++)
 					<input type="hidden" name="order[]" value="<?php echo $row->ordering; ?>" />
 				</td>
 			</tr>
-<?php
-	$k = 1 - $k;
-}
-?>
+			<?php
+			$k = 1 - $k;
+		}
+		?>
 		</tbody>
 	</table>
